mime: cast to int (*char) argument of isspace()
to get rid from warnings during build with clang

+++ src/mime.c
@@ -115,7 +115,7 @@ char *get_mime_type(char *filename)
 		}
 		ptr++;
 
-		while (!isspace(*ptr))
+		while (!isspace((int)*ptr))
 		{
 			buffer[i] = *ptr;
 			ptr++;
